refactoring people bloc
